fix: remove getFavoriteCounts
